Completed single layer denoising autoencoder.
The network is 784 -> 588 -> 788 and uses Leaky ReLUs with nesterov momentum for updates.

There's a lot of diagnostics included. 4 charts: loss per epoch, histogram of hidden layer activations, visualization of hidden layer weights, and visualization of reconstruction.

The learning rate is changed automatically if the weight/update ratio is not within an epsilon of 1e-3 (as per cs231).